chore(LicenseManager): refactor to use input assertion abstraction
